requirements.txt
# Core orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq
langchain-huggingface

# Retrieval
ddgs
tavily-python
sentence-transformers

# Eval
ragas
datasets

# UI
gradio>=4.0.0

# Utilities
python-dotenv
pandas
tqdm
networkx
matplotlib